fix a/v sync!!!
Created lifeAI Player that plays audio/video and burns in
the subtitles.

This is a solution that works since we are AI and the actual
frame to audio/text really doesn't matter too much here ;)

video encoding / audio and streaming the assets will be the
next step to allow direct output to twitch.